Changed: Moved the 'debug' module CLI into 'debug_format' module.
