+++ projects/tools/dxt-lossless-transform-cli/Cargo.toml
@@ -11,15 +11,15 @@ multithreaded = ["rayon"]
 
 # Debugging and Research Tools/Utilities for BC7
 # Currently BC7
-debug-bc7 = ["debug"]
+debug-bc7 = ["debug-format"]
 # Debugging BC1 behaviour
-debug-bc1 = ["debug", "dxt-lossless-transform-bc1/experimental"]
+debug-bc1 = ["debug-format", "dxt-lossless-transform-bc1/experimental"]
 # Debugging BC2 behaviour
-debug-bc2 = ["debug", "dxt-lossless-transform-bc2"]
+debug-bc2 = ["debug-format", "dxt-lossless-transform-bc2"]
 # Debugging endianness compatibility
-debug-endian = ["debug", "tempfile"]
-# Common debugging utilities
-debug = ["zstd-sys", "thiserror-no-std", "xxhash-rust", "bincode", "serde", "dirs", "derive-enum-all-values", "lossless-transform-utils", "dxt-lossless-transform-dds/debug-block-extraction", "dxt-lossless-transform-file-formats-debug"]
+debug-endian = ["tempfile"]
+# Common debugging utilities for file format analysis
+debug-format = ["zstd-sys", "thiserror-no-std", "xxhash-rust", "bincode", "serde", "dirs", "derive-enum-all-values", "lossless-transform-utils", "dxt-lossless-transform-dds/debug-block-extraction", "dxt-lossless-transform-file-formats-debug"]
 
 nightly = ["dxt-lossless-transform-bc1/nightly", "dxt-lossless-transform-common/nightly", 
            "dxt-lossless-transform-api-common/nightly", "safe-allocator-api/nightly"]
