+++ projects/tools/dxt-lossless-transform-cli/src/debug_format/benchmark_common.rs
@@ -3,7 +3,7 @@
 //! This module provides shared data structures, utilities, and benchmarking functions that can be
 //! reused across different BC format implementations for performance analysis.
 
-use crate::debug::calc_compression_stats_common::get_filename;
+use crate::debug_format::calc_compression_stats_common::get_filename;
 use crate::util::Throughput;
 use core::{f64, fmt::Debug};
 use std::{
