+++ projects/tools/dxt-lossless-transform-cli/src/debug_format/calc_compression_stats_common.rs
