+++ projects/tools/dxt-lossless-transform-cli/src/debug_format/compressed_data_cache.rs
