+++ projects/tools/dxt-lossless-transform-cli/src/debug_format/compression_size_cache.rs
